backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
slowapi==0.1.9
anthropic==0.39.0
openai==1.51.0
google-generativeai==0.8.3
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
